pyproject.toml
[project]
name = "hikari-crescent"
version = "0"
description = "🌙 A command handler for Hikari that keeps your project neat and tidy."
authors = [
    { name = "Lunarmagpie", email = "bambolambo0@gmail.com" },
    { name = "circuitsacul", email = "circuitsacul@icloud.com" },
]
requires-python = ">=3.10"
readme = "README.md"
license = "MPL-2.0"
maintainers = [
    { name = "Lunarmagpie", email = "bambolambo0@gmail.com" },
    { name = "circuitsacul", email = "circuitsacul@icloud.com" },
]
keywords = [
    "discord",
    "hikari",
    "command handler",
    "slash commands",
    "application commands",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = ["hikari>=2.3.0,<3"]

[project.optional-dependencies]
i18n = ["python-i18n>=0.2"]
cron = [
    "croniter>=5.0.0,<6",
    "types-croniter>=5.0.0,<6",
]
cooldowns = ["floodgate-rs>=0.1.1,<0.2"]

[project.urls]
Homepage = "https://github.com/hikari-crescent/hikari-crescent"
Repository = "https://github.com/hikari-crescent/hikari-crescent"
Documentation = "https://hikari-crescent.github.io/hikari-crescent/crescent.html"

[dependency-groups]
dev = [
    "mypy>=1.4.0,<2",
    "pyright>=1.1.315,<2",
    "ruff>=0.10.0,<0.11",
    "codespell>=2.2.5,<3",
    "mkdocs-material[imaging]>=9.5.39,<10",
    "mkdocs>=1.6.1,<2",
    "mkdocstrings-python>=1.11.1,<2",
    "griffe-inherited-docstrings>=1.0.0,<2",
    "pytest>=8.2.2,<9",
    "pytest-asyncio>=0.23.7,<0.24",
    "python-dotenv>=1.0.0,<2",
    "hikari[server]",
]

[tool.uv.build-backend]
module-name = ["crescent"]
module-root = ""

[build-system]
requires = ["uv_build>=0.9.9,<0.10.0"]
build-backend = "uv_build"

[tool.mypy]
python_version = "3.10"
strict = true
warn_unused_configs = true
warn_return_any = true
warn_redundant_casts = true
namespace_packages = true
warn_unused_ignores = false

[tool.pyright]
include = ["crescent", "examples"]
exclude = ["tests"]
pythonVersion = "3.10"
typeCheckingMode = "strict"
reportPrivateUsage = false
reportImportCycles = false
reportIncompatibleMethodOverride = false
reportWildcardImportFromLibrary = false

[tool.ruff]
target-version = "py310"
lint.select = ["E", "W", "F", "RUF"]
line-length = 99

[tool.ruff.lint.pyflakes]
extend-generics = [
    "crescent.Plugin"
]

[tool.pytest.ini_options]
asyncio_mode = 'strict'
